Restructure compose theming sources
