+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Yetanotherreddit">
         <activity
-            android:name=".feature.feed.FeedActivity"
+            android:name=".feature.feed.ui.FeedActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Yetanotherreddit.NoActionBar">
