+++ app/src/main/java/com/olderwold/reddit/feature/feed/ui/FeedActivity.kt
@@ -1,4 +1,4 @@
-package com.olderwold.reddit.feature.feed
+package com.olderwold.reddit.feature.feed.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -6,8 +6,8 @@ import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
+import com.olderwold.reddit.feature.feed.ui.theme.YetanotherredditTheme
 import com.olderwold.reddit.feature.web.WebPage
-import com.olderwold.reddit.ui.theme.YetanotherredditTheme
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
