+++ app/src/main/java/com/olderwold/reddit/feature/feed/ui/FeedList.kt
@@ -1,4 +1,4 @@
-package com.olderwold.reddit.feature.feed
+package com.olderwold.reddit.feature.feed.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
