+++ app/src/main/java/com/olderwold/reddit/feature/feed/ui/FeedPagingSource.kt
@@ -1,4 +1,4 @@
-package com.olderwold.reddit.feature.feed
+package com.olderwold.reddit.feature.feed.ui
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
