+++ app/src/main/java/com/olderwold/reddit/feature/feed/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.olderwold.reddit.ui.theme
+package com.olderwold.reddit.feature.feed.ui.theme
 
 import androidx.compose.ui.graphics.Color
